+++ CHANGELOG.md
@@ -11,6 +11,8 @@ If migrating from Carbon:
 * Rename usages of `Set-CEnvironmentVariable` with `Set-CEnvVariable`.
 * `Remove-CEnvVariable` now writes an error if the environment variable to remove does not exist. Add `-ErrorAction
   Ignore` to ignore any errors
+* `Set-CEnvVariable` now writes an information message when setting an environment variable. The message includes the
+  value. To omit the value from the messages, use the new `Sensitive` switch.
 
 ### Added
 
@@ -22,3 +24,5 @@ If migrating from Carbon:
 
 * `Remove-CEnvVariable` writes an error if an environment variable doesn't exist.
 * `Remove-CEnvVariable` writes an information message for each environment variable it deletes.
+* `Set-CEnvVariable` writes an information message for each environment variable it sets, including the variable's
+  value. Use the `Sensitive` switch to omit the value from the information message.